remove warnings

+++ Genix/src/Genix/ImGui/ImGuiLayer.cpp
@@ -63,7 +63,7 @@ void ImGuiLayer::OnUpdate()
 {
 	ImGuiIO& io = ImGui::GetIO();
 	Application& app = Application::Get();
-	io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+	io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
 	float time = (float)glfwGetTime();
 	io.DeltaTime = m_Time > 0.0f ? (time - m_Time) : (1.0f / 60.0f);
@@ -161,7 +161,7 @@ bool ImGuiLayer::OnKeyTypedEvent(KeyTypedEvent& e)
 bool ImGuiLayer::OnWindowResizeEvent(WindowResizeEvent& e)
 {
 	ImGuiIO& io = ImGui::GetIO();
-	io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+	io.DisplaySize = ImVec2((float)e.GetWidth(), (float)e.GetHeight());
 	io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 	glViewport(0, 0, e.GetWidth(), e.GetHeight());
